--- src/main/kotlin/astminer/common/TreeUtil.kt
@@ -1,2 +1,0 @@
-package astminer.common
-
